requirements.txt
requests_html
pillow
